requirements.txt
pdf2image~=1.16.2
pypdf2~=3.0.1
Pillow~=9.4.0
python-docx~=0.8.11